refactor: Optimize codebase by restructuring files
- Reorganize API calls into separate files for better organization and maintainability.
- Split functions into distinct files , enhancing code modularity and clarity.
