+++ index.html
@@ -131,7 +131,7 @@
       <!-- Append the listOfVids here -->
     </div>
   </div>
-  <script src="/client.js"></script>
+  <script type="module" src="/client.js"></script>
 </body>
 
 </html>
